Extract ScreenSharingPlayers component
